requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
jinja2
transformers
torch
openai
python-dotenv